Try to repro on a smaller dataset
It no longer repros.
